% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsim.R
\name{bsim}
\alias{bsim}
\title{bsim}
\usage{
bsim(
  s = seq(0, 1000, by = 100),
  bc = 0.5,
  df = c("G1", "G2", "G3", "G4", "G5", "G6", "G7", "G8"),
  v = 1200,
  sh = 1.6,
  angle = 0,
  zero = 100,
  windspeed = 0,
  windangle = 0
)
}
\arguments{
\item{s}{a}

\item{bc}{a}

\item{df}{drag function: G1, G2, G3, G4, G5, G6, G7, or G8}

\item{v}{a}

\item{sh}{a}

\item{angle}{a}

\item{zero}{a}

\item{windspeed}{a}

\item{windangle}{a}
}
\description{
A short description
}
